API orderDetail

+++ shopapp-backend/shopapp/src/main/java/com/project/shopapp/dtos/OrderDetailDTO.java
@@ -28,7 +28,7 @@ public class OrderDetailDTO {
 
     @JsonProperty("total_money")
     @Min(value = 0, message = "total money >=0")
-    private Long totalMoney;
+    private Float totalMoney;
 
     private String color;
 }
